app/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==2.1.15
filelock==3.7.1
greenlet==1.1.2
idna==3.3
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==6.0.4
isodate==0.6.1
lxml==4.8.0
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.2.0
platformdirs==2.5.1
python-decouple
psycopg2-binary==2.8.6
pymssql==2.2.7
python-dateutil==2.8.2
pytz==2022.1
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.10
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
simplejson==3.17.6
six==1.16.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.35
sqlparse==0.4.2
typing-extensions==4.5.0
urllib3==1.26.9
zeep==4.1.0
zipp==3.15.0
